KL-139/rename: move file
- 파일 이동 loginModal, SearchModal
