+++ src/components/Navbar/components/LoginModal.jsx
@@ -2,9 +2,9 @@ import React from 'react'
 import styled from 'styled-components'
 import { Modal, Button, ConfigProvider } from 'antd'
 import { useShallow } from 'zustand/react/shallow'
-import { KakaoLogo, NaverLogo } from '../../images/logos'
-import { useModalStore } from './stores/navbarStores'
-import theme from '../../styles/theme'
+import { KakaoLogo, NaverLogo } from '../../../images/logos'
+import { useModalStore } from '../stores/navbarStores'
+import theme from '../../../styles/theme'
 
 const ModalTheme = {
   components: {
